Move ads/adsense.* and ads/doubleclick.* to ads/google/ Add ads/google/utils.js, as a place for code to be shared by adsense.js and doubleclick.js.
