+++ 3p/integration.js
@@ -28,11 +28,11 @@ import {a9} from '../ads/a9';
 import {adblade, industrybrains} from '../ads/adblade';
 import {adform} from '../ads/adform';
 import {adreactor} from '../ads/adreactor';
-import {adsense} from '../ads/adsense';
+import {adsense} from '../ads/google/adsense';
 import {adtech} from '../ads/adtech';
 import {plista} from '../ads/plista';
 import {criteo} from '../ads/criteo';
-import {doubleclick} from '../ads/doubleclick';
+import {doubleclick} from '../ads/google/doubleclick';
 import {dotandads} from '../ads/dotandads';
 import {endsWith} from '../src/string';
 import {facebook} from './facebook';
