+++ ads/google/utils.js
@@ -1,0 +1,29 @@
+/**
+ * Copyright 2016 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * @param {!Window} global
+ * @return {number}
+ */
+export function getCorrelator(global) {
+  const clientId = global.context.clientId;
+  const pageViewId = global.context.pageViewId;
+  if (global.context.clientId) {
+    return pageViewId + (clientId.replace(/\D/g, '') % 1e6) * 1e6;
+  } else {
+    return pageViewId;
+  }
+}
