+++ builtins/amp-ad.md
@@ -58,11 +58,11 @@ resources in AMP. It requires a `type` argument that select what ad network is d
 - [Adblade](../ads/adblade.md)
 - [Adform](../ads/adform.md)
 - [AdReactor](../ads/adreactor.md)
-- [AdSense](../ads/adsense.md)
+- [AdSense](../ads/google/adsense.md)
 - [AdTech](../ads/adtech.md)
 - [Criteo](../ads/criteo.md)
 - [Dot and Media](../ads/dotandads.md)
-- [Doubleclick](../ads/doubleclick.md)
+- [Doubleclick](../ads/google/doubleclick.md)
 - [Flite](../ads/flite.md)
 - [Improve Digital](../ads/improvedigital.md)
 - [Industrybrains](../ads/industrybrains.md)
